Fix compiler warnings related to lossy conversion and new name for ROOTWriter (#46)
* Fix compiler warnings related to double - float

* Remove warning that no longer applies

* Switch to ROOTWriter from deprecated ROOTFrameWriter

---------

+++ k4EDM4hep2LcioConv/include/k4EDM4hep2LcioConv/k4EDM4hep2LcioConv.ipp
@@ -672,10 +672,9 @@ namespace EDM4hep2LCIOConv {
         lcio_mcp->setEndpoint(endpoint);
         double momentum[3] = {edm_mcp.getMomentum()[0], edm_mcp.getMomentum()[1], edm_mcp.getMomentum()[2]};
         lcio_mcp->setMomentum(momentum);
-        float momentumEndpoint[3] = {
+        double momentumEndpoint[3] = {
           edm_mcp.getMomentumAtEndpoint()[0], edm_mcp.getMomentumAtEndpoint()[1], edm_mcp.getMomentumAtEndpoint()[2]};
         lcio_mcp->setMomentumAtEndpoint(momentumEndpoint);
-#warning "double to float"
         lcio_mcp->setMass(edm_mcp.getMass());
         lcio_mcp->setCharge(edm_mcp.getCharge());
         float spin[3] = {edm_mcp.getSpin()[0], edm_mcp.getSpin()[1], edm_mcp.getSpin()[2]};
